関数をmodelsディレクトリに集約
