--- public/view/get-all-schedules.html
@@ -1,6 +1,0 @@
-{{define "get-all-schedules"}}
-{{range .}}
-{{.ID}}
-{{.EndHour}}
-{{end}}
-{{end}}
